Update reference element for cursor selection when root element updates
