+++ packages/liveblocks-react-lexical/src/liveblocks-plugin-provider.tsx
@@ -15,6 +15,7 @@ import React, {
   useRef,
   useState,
 } from "react";
+import { useSyncExternalStore } from "use-sync-external-store/shim/index.js";
 import { Doc } from "yjs";
 
 import { CommentPluginProvider } from "./comments/comment-plugin-provider";
@@ -218,14 +219,14 @@ export const LiveblocksPlugin = ({
     collabContext.name = username || "";
   }, [collabContext, username]);
 
-  useLayoutEffect(() => {
-    const editable = editor.getRootElement();
-    if (editable === null) return;
+  const root = useRootElement();
 
+  useLayoutEffect(() => {
+    if (root === null) return;
     setReference({
-      getBoundingClientRect: () => editable.getBoundingClientRect(),
+      getBoundingClientRect: () => root.getBoundingClientRect(),
     });
-  }, [setReference, editor]);
+  }, [setReference, root]);
 
   const handleFloatingRef = useCallback(
     (node: HTMLDivElement) => {
@@ -266,3 +267,20 @@ export const LiveblocksPlugin = ({
     </>
   );
 };
+
+function useRootElement(): HTMLElement | null {
+  const [editor] = useLexicalComposerContext();
+
+  const subscribe = useCallback(
+    (onStoreChange: () => void) => {
+      return editor.registerRootListener(onStoreChange);
+    },
+    [editor]
+  );
+
+  const getSnapshot = useCallback(() => {
+    return editor.getRootElement();
+  }, [editor]);
+
+  return useSyncExternalStore(subscribe, getSnapshot, getSnapshot);
+}
